Moved App outside and pages in its own folder
