+++ src/pages/Playground.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Playground from './App';
+import Playground from '../App';
 
 describe('App', () => {
   it('should not throw on render', () => {
